Use TextureView IDs to query VkFramebuffer cache
Previously the framebuffer cache was using WeakPtrs to TextureViews to
generate the hash, but this turned out to be fairly expensive as we
needed to promote the WeakPtr to a Ref for each attachment every time
we did a cache lookup. Unfortunately this made the framebuffer cache
almost as expensive as simply re-creating the framebuffers.

By switching to use a Device-unique ID for texture views during the
cache lookup we avoid all of that overhead.

Bug: 416088623
Change-Id: I8db50bbbb9f97d10bcbd148a5561fbd7b02b60ca
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/255615
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>