Dawn: Add integrity check for BlobCache under toggle
This CL add integrity check in StoreInternal/LoadInternal of BlobCache
using SHA3 to transparently verify the payload read from cache, if the
hash result doesn't match it would be treated as cache miss and record
in the histogram.

The validation is controlled by a new added toggle
BlobCacheHashValidation, which toggle is enabled by default for now.

Bug: 427397692, 429938352
Change-Id: I890158f1121d347c548cc5d0e5a3e223f9b55c5d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249414
Commit-Queue: Zhaoming Jiang <zhaoming.jiang@microsoft.com>
Reviewed-by: Loko Kung <lokokung@google.com>