[hlsl] Remove list of entry points from writer output
The backend assumes that it is only generating a single entry point,
so we do not need to return a list. This simplifies the code at the
callsites in Tint that validate the generated HLSL.

This also removes the last build dependency on the `wgsl` directory
from `hlsl`, by switching over to the IR version of the PipelineStage
enum.

Change-Id: I0e63e76e9ce5d7fabe42745a493006b67b1e74fb
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/250820
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>