[spirv-reader][ir] Fix propagation from if block
In the case of a `continuing` block pulling a value from the loop body,
when an `if` block jumps to the continuing we didn't know to pass the
extra parameters. We have to pass the continuing PHI params through in
the case we don't know what the parameters should be.

Bug: 435621075
Change-Id: I57f8a5dcf96457eac8d9c0f7351907211f3f91b2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256054
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>