[spirv-reader][ir] Fix constant-id check for NonUniform calls.
The original check that the `value` was constant was a) too broadly
applied and b) the wrong operand to check. This Cl constrains the check
to just the Broadcast, QuadBroadcast and QuadSwap instructions and
corrects the check to use the `Invocation Id` instead of the `Value`.

Fixed: 432807736
Change-Id: I643d3e44a77aa65ed8f9ce26c1fb9ade2c5b8fd6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256374
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>