Stop requiring Device for AttachmentState creation
AttachmentStates inherited from ObjectBase, which requires them to hold
a Ref to a Device, but they never actually used it outside of one
feature check in a debug message. This change switches them to inherit
from RefCounted, which avoids the need to track the Device. This in turn
speeds up the creation of the AttachmentStates Blueprints used to query
the cached AttachmentStates.

Bug: 436878521
Change-Id: Iefff5d4e12141b7072fd244221ba113456412ac0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256274
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>