[spirv-reader] Do not change access mode used for type map for handles
The access mode that we use to key into the type map needs to use the
actual access mode that we are going to use for the image type,
otherwise we will have a collision when the same SPIR-V type is used
in two different variables with two different access mode.

We already handle setting the access mode for image types when we
actually emit the type, so we don't need to change anything for the
type map key.

Fixed: 437093642
Change-Id: I90b6e383c6ad0bc563fcea43ee698f7a81723254
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256517
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>