[dawn][d3d12] Fix texture component swizzle for stencil formats
The existing logic for composing the texture component swizzle for
stencil formats was incorrect. It did not properly apply the existing
texture view's swizzle. This patch refactors the code to use a new
helper function, `ComposeSwizzle`, that correctly combines the base
texture swizzle with the view's swizzle. This ensures that the texture
components are correctly remapped when a texture view is created for a
stencil format.

Bug: 414312052
Change-Id: I8ffe83cad83e108bb7074138e62c3215f41ef67e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256414
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Gregg Tavares <gman@chromium.org>