[tint] Always expose SpirvToWgsl
An external user that includes `tint.h` will not have defined the
macros needed to enable this function. We should not require that
external users always build their code with the same macros that Tint
was built with.

Instead, this change always exposes SpirvToWgsl, which will then
return an error if Tint itself was not built with the required macros.

Change-Id: I154a3f91aa5bedb690b2d98139c44cf4c569e796
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256576
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>