Remove the syntax tree printer.
With the conversion to IR, there are really only two places we have AST
at this point, coming out of the parser and into the IR and coming out
of IR for WGSL printing.

With the removal of transformations on the AST the syntax tree parser
has much less utility. Removing it.

Change-Id: I1a324f783a44fe4e5f60558f3f176221ad4678c7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256694
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>