Kotlin: Bundle sources with the library.
Desipte `sourceSets` in `build.gradle` we weren't bundling sources.

This refers to the dev-facing Kotlin, and it's immensely useful to get
this info because it includes defaults for structures and method
parameters.

It's 72KB currently which compresses to 50KB.

Bug: 437492284
Test: ./gradlew publishToMavenLocal`, then look for
`~/.m2/repository/net/android/webgpu/1.1.34/webgpu-1.1.34-sources.jar`

Change-Id: I51e9fadd8d6095f473ab693e9453ca718d6c1276
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256615
Commit-Queue: Jim Blackler <jimblackler@google.com>
Reviewed-by: Nikita Gupta <nikigupta@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>