Patch 1/8: Implement FSTestOSWrapper foundations
Defines a testing/fstest based implementation of the OSWrapper
interface, FSTestOSWrapper. This is a stub implementation with minimal
amount of the constructor and utility functions needed, otherwise
methods panic with a message about not being implemented.

Implemented all of the various testing helpers that will be needed in
future CLs in fstestoswrapper_test.go

Renamed the test names in memmaposwrapper_test.go to be prefixed with
the specific implementation being tested, so that testing logs are
easier to read.

Bug: 436025865
Change-Id: I17dda6488b9d5c88049f09087105cf2ed34cc07d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256435
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@google.com>