Patch 2/8: Implement core read/write operations for FSTestOSWrapper
Implements ReadFile() and WriteFile() and adds testing for them, since
they are needed to test each other, and pretty much all of the other
functions.

Also implements MkdirAll(), since it is needed in one of the testing
helpers, but does not include tests for it, since its tests require
other functions to be implemented. There is effectively a cycle in the
dependency graph between functions and their tests here that needs to
be broken to avoid a large monolithic patch.

Bug: 436025865
Change-Id: I109dd0a071b23a514120052549c800a007c64ee8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256455
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>