Enable Dawn standalone CQ builder
Enables dawn-cq-linux-x64-rel to be added on (almost) all CLs and block
CL submission on failure. This is currently the same as the
dawn-try-linux-x64-sws-rel manual-only trybot, but will have additional
tested configs added to it as more Dawn standalone builders are moved to
the gn_v2 recipe.

Bug: 385317083
Change-Id: I8dcd197be1a20d97635011c1c045bd0b7dcb461b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/257414
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>