Remove dawn-cq-linux-x64-rel from CQ
Temporarily removes the recently added dawn-cq-linux-x64-rel from the
Dawn CQ. There is currently a bug in the recipe that is causing test
failures to not surface as build failures. The CI equivalent should
currently be red, so we need to remove this from the CQ until the
recipe is fixed and the CI builder is green.

Bug: 385317083
Change-Id: I063a1c48d82be88daaaa81455da0cd9c30752d6c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/257674
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>