[dawn][vk] Enable bindless using "descriptor indexing"
Vulkan's equivalent of bindless is called "descriptor indexing" and a
soup of features and limits. Add support for this Vulkan feature and
gate WebGPU's bindless on a bunch of features and limits.

Note that "robustBufferAccessUpdateAfterBind" is a device property that
has implication for robustness when bindless is available. It will be
handled in a future CL and is not handled here.

Bug: 435251399
Change-Id: I9e77026c53b83dc72b30532c12541bd6193871d1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256936
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>