[dawn][vk] Handle robustBufferAccessUpdateAfterBind.
When this property is false, all forms of buffer robustness are
disallowed, except pipelineRobustness's vertexInputs robustness.

This CL adds:

 - Detection of the VK_EXT_pipeline_robustness extension (promoted to
   Vulkan 1.4).
 - Gates bindless on robustBufferAccessUpdateAfterBind or
   pipelineRobustness being present.
 - Adds logic to disable robustness on buffers (and not rely on it) when
   bindless is enabled and robustBufferAccessUpdateAfterBind is false.
 - Adds logic to request when creating a pipeline that robust access
   still happen on vertexInputs.

Altogether this changes aim at supporting bindless anywhere we can
enable robustness on vertex inputs (because we don't want to emulate
that if we can avoid it), while taking advantage of as much of the
driver's robustness support as possible (like we currently do).

Bug: 435251399
Change-Id: I2164e3396bcbdc2f4c215ffe1568147c2d6aa062
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/257156
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>