[tint][cmake] Use build protoc instead of system installed version
The existing cmake rule will use the first protoc that shows up in your
path, so will likely default to the system installed version of protoc,
which may not be compatible with the version of libprotobuf in our repo.
Additionally this can lead to build failures when there is no installed
protoc, and the output path for the build is not in the path.

This also include a small fix for '"--${ARGS_LANGUAGE}_out', since
protoc uses '--<flag>=<arg>' or '-<flag><arg>' format for its CLI, not
'--<flag> <arg>'. And it was using the wrong variable name.

Fixes: 435357782
Change-Id: I20c921599caf5ac4b5c57545fb61886463092a13
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/257894
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: James Price <jrprice@google.com>