[dawn][vk] Freshen up comments in vulkan::{BGL, DescriptorSetAllocator}
These were outdated, had TODOs with too few details and associated with
people instead of issues, etc.

Also makes `DescriptorSetAllocator` a `RefCounted` as it doesn't need to
be `ObjectBase` with the error monad mechanism.

Also pass in the VkDescriptorSetLayout instead of the BindGroupLayout to
DescriptorSetAllocator::Allocate to make it more clear exactly what will
be used.

Bug: 439522242
Change-Id: Iac766c4f0368dadab4ee7db67deda63b1023fbca
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/257754
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>