[dawn][native][d3d12] Force higher minimums for maxDynamic* limits
Adding maxImmediateSize to D3D12 took space away from
maxDynamicUniformBuffersPerPipelineLayout and
maxDynamicStorageBuffersPerPipelineLayout, which is a regression.
Set those first, so that allocating maxImmediateSize can't take away
from them.

This should be safe for all machines because kMaxRootSignatureSize and
the default values for the limits are all constants, so all this
allocation logic has the same logic on all machines.

Fixed: 440381283
Change-Id: I679481bf6a31cda73c117d5b7c9518b59f9008b6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/259055
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>