[dawn][native] Require BGL dynamic array even for size 0 BG dynamic array
Prior to this CL, a special case in the validation allowed a BG with a
dynamic array size of 0 to be compatible with a BGL without a dynamic
array. This turned out to cause a few issues. Disallow it for now and
revisit later depending on what the spec / webgpu.h decides.

Bug: 435317394
Change-Id: Ieb4f6f14d47a4ff6c662275c97f8a495c1b2ae6a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/258656
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>