[dawn][wire] Crash client when using objects from another wire
This is of course the incorrect behavior (we are supposed to produce a
"[object] cannot be used with [device]" validation error), but it does
prevent old client objects from misassociating with new server objects.
This avoids any possible renderer-process security issue when that
happens (though none has been identified). (There are not believed to be
any GPU-process problems here since that interface is already fuzzed.)

This crash can only happen after the GPU process has already crashed.
This is of course possible to do intentionally, but the most likely
situation where this happens in the wild is in an application that has
*some* device loss recovery support but it's not well tested, or it
mixes up objects in a way that doesn't break the application.

It's also technically possible for an application to try to rely on
push+popErrorScope to check whether objects are from the same device.
That would work fine in most cases (e.g. device.destroy(), which would
be the typical way for applications to test their recovery behavior), it
would just not work in cases where the GPU process has crashed.

Fixed: 435630456
Bug: 440387003
Change-Id: I73a2d816358bbfe9840cfecfe954c3d3c1b34bcf
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/256757
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>