Run `gn format` directly.
Currently the generation script runs the `gn.py` script through
depot_tools using `vpython3`. Instead. run the `gn` which is pulled in
through the DEPS file directly.

Note, this will only work on a linux, mac or windows machine as we only
pull `gn` for those versions. If we find we have developers on other
platforms, we may want to conditionalize this and fall back to the
python `gn.py` script to find `gn`.

This cuts about 4s off the gen time on my M3.

Change-Id: Ifc6201e8088bc9f65da573f633d8bf9574160bb8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/259256
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>