[dawn][native] Add a DynamicArray metadata buffer internal binding.
Dynamic binding arrays need metadata alongside the actual bindings in
the array so that code generated from WGSL can check for OOB and that
the requested type matches what's in the array.

Add a temporary validation check that external textures are not allowed
in the same BGL as a dynamic binding array as it causes issues with the
allocation of BindingNumber for expanded entries.

Bug: 435317394, 42240282
Change-Id: I11df49ad6b781ac2989a7a3b55582c234d7a59a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/258694
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>