[dawn][native] Add Texture::Pin/Unpin frontend validation.
The bindless extensions need a way to reduce the cost of memory barrier
tracking and usage scope validation. Usage pinning is the chosen method
for the prototyping in Dawn: it forces the texture to stay in a specific
usage until unpinned (or re-pinned). Backends no longer need to track
the texture as they can emit the necessary memory barrier when Pin() is
called.

Note that part of the validation for Pin() is missing because at submit
time we should check that all non-bindless usages of the texture are in
the pinned set of usages, but doing so could cause performance
regressions so it is skipped for now.

Bug: 435317394
Change-Id: Iaed7e5c77d35568ff8d3978d32f09ecad7f138c2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/258995
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>