Add a length check to `hasBinding`.
When we do the `hasBinding` check make sure that we're within the array
size. If we're outside, we'll get robustness kicking in and give us a
value from the array which isn't what we want. So, verify we're within
the buffer size before trying to check the kind at the given index.

Bug: 441986708
Change-Id: Id9a24f929d2e5b0aa768e708e1152951c169c5a3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/259894
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>