[tint] Add missing includes
When Tint is built with only the SPIR-V reader and WGSL writer enabled
(e.g. because it's being used for SPIR-V to WGSL conversion and
nothing else), we need to explicitly include the ir::Module header.

We also need `tint::Program` when the WGSL writer is enabled, for the
program printer.

Change-Id: I47f0f48dd173c4f16bffb7e5a0548193ea62b634
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/260794
Reviewed-by: Peter McNeeley <petermcneeley@google.com>
Commit-Queue: James Price <jrprice@google.com>