[dawn] Ensure release bounds checks in ityp::*
Add death tests to ensure out-of-bounds accesses in all the ityp::*
containers crashes in release.

Both libc++ and absl are supposed to be configured to guarantee this, so
we don't need our own checks for those. (MSVC is not configured this
way, and I don't think we test libstdc++. Both are capable though.)

Fix ityp::array/vector/stack_vec by just removing the DAWN_ASSERT. That
was telling the compiler to assume its condition is true, causing it to
elide the bounds checks inside libc++ (that are supposed to always be
enabled for safety).

Fix ityp::span by changing the DAWN_ASSERT to a DAWN_CHECK. This should
really be changed to be based on std::span but that should be done
separately.

Bug: 442444724, 442860471
Change-Id: I7c37fc9c382b840a09df0b00dd782a5cdb80de6e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/260462
Reviewed-by: Gregg Tavares <gman@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>