[dawn][native] Cap maxImmediateSize at 32
maxImmediateSize is, for now, limited to 32 on at least all D3D12
devices, so we probably shouldn't expose higher values until we have a
way to do so more widely. We can increase it if we make more space in
the D3D12 root descriptor.

Bug: 366291600
Change-Id: I8448bc7d9387ae6a9d3452cee1c88c40080c3038
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/259094
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Shaobo Yan <shaoboyan@microsoft.com>