[setup-build] fix 'cmake install' scenario
The 'install' option configures Dawn to be built
as a monolithic library, either as shared or static.
This is incompatible with BUILD_SHARED_LIBS.

Rearrange the logic so BUILD_SHARED_LIBS is only turned
on when *not* installing.

Signed-off-by: dneto <dneto@google.com>
Change-Id: I6a6a6964837e57a5a6a64da5c99878e1922bb127
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/260796
Commit-Queue: Peter McNeeley <petermcneeley@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Peter McNeeley <petermcneeley@google.com>