Make wgpu::Buffer::GetMapState() no autolock
The buffer state is tracked in an atomic so it's safe to check without
acquiring the device lock. There are no guarantees the buffer state
doesn't change on another thread immediately after GetMapState() returns
regardless so it's up to the application to properly synchronize buffer
access.

Bug: 425472913
Change-Id: Ice847384e0139bf6446eb102d39d789ea9e945c5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261274
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>