Lower name truncation threshold
Lowers the threshold at which we automatically truncate the parameters
included in Dawn test names by 10 characters. This is a speculative fix
for dawn_end2end_tests failing to upload results to ResultDB on certain
Android Desktop configurations. The one known case where this is
happening results in a test name that is 258 characters long, which is
over the 256 character limit.

Bug: 413427416
Change-Id: Ibefb0cb4355a6dbffe27b3b2c9fdbbca903822e1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261316
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>