Handle invalid from type in load/store vector element.
The vector element validation was checking that the From/To value was a
vector pointer, but it would not exit if that was not the case. This
then caused later code to fail when it asserted that fact. Update the
code to return immediately when a bad From/To type is provided.

Fixed: 444353801
Change-Id: I3fa61d11b917c17b545ad35d8b9c0940a256749f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261514
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>