Refactor test name truncation
Refactors the logic used for determining if test names need to be
truncated in order to fit within the 256 character limit for ResultDB.

Previously, we checked if the test class + some fixed value + the
sanitized name was less than some arbitrary number of characters. This
was problematic since it did not cleanly take into account any prefixes
added to the test class or test names that were significantly longer
than the fixed value we used, leading to cases where tests were not
truncated when they should be.

Now, we keep track of the longest test name we've seen in the wild and
truncate if that + the sanitized parameters are over a threshold.

Bug: 413427416
Change-Id: I097f7a9f58ab5f2d74f61d0750e1fb60c3fa167f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261440
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>