[tint][val] Check that the type for @location is valid
Enforces that it is only attached to numeric scalars or vectors.

Also adds a capability to allow location on structs, matrices, and
arrays when they have only numeric elements, so that the 'from SPIRV'
path can handle inputs with this construction, instead of immediately
bailing.

The duplicate annotation test needed to be updated, since it
previously depended on location being attached to a struct directly
was valid, which is no longer allowed.

Fixes: 440157916
Change-Id: I73c4e2eeee5d2d93b7d3d6948889966b82f866dd
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261354
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>