[tint][val] Restrict resource_binding usage
Adds a check that resource_binding, an internal IR type, isn't used
and an associated capability to allow it. This capability is enabled
when lowering to core IR and raising to SPIRV until
transform::ResourceBinding is run.

This implicitly bans the resource_binding specific builtins, since any
code that uses them will either include the disallowed type, or
the matcher will fail.

This fixes issues with the fuzzers generating inputs with
resource_binding causing failures in transforms that are not meant to
support it.

Fixes: 444567409
Change-Id: I72d4ac6376fdae64cf49a3f0f10ccd098ed82c49
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262234
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>