Fetch third_party/node via CIPD
Switches from using a GCS hook for third_party/node to a CIPD
dependency. This is specifically being added to unblock some work
related to running WebGPU CTS tests on all platforms in the gn_v2
recipe, but should also be a quality of life improvement for managing
this dependency.

The requested version remains unchanged, but should be easy to change in
the future. The version that is currently used is >2 years old at this
point.

NOTE: Due to this switch, the paths on Linux and Mac have changed
slightly since there is no longer an intermediate directory between
//third_party/node/ and the actual package contents. However, because
of caching on the bots, the old GCS-fetched data may persist for up to a
week after this lands. As such, it is possible that failures will start
occurring at a later point if the old paths are still used somewhere. If
this occurs, it is likely better to fix the problematic paths instead of
reverting this.

Similarly, the GCS-fetched data will likely remain in local checkouts.
This is not expected to cause any issues, but users can delete the
contents of third_party/node/ and re-run "gclient sync" to remove it if
desired. "gclient sync -D" does not appear to clean this up
automatically.

Bug: 441328362
Change-Id: I6fb587804a49cb63db3aa95bc23a16d763b89a33
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262334
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>