[spirv-reader] Pass textures and samplers as non-pointers
When textures/samplers are passed in SPIR-V the come as pointers, in
WGSL we need them to be non-pointer handle types. Fix up the pointers
when creating the function and load as needed at the call sites.

Fixed: 441874372
Change-Id: I444886f40fc0d30f5eceaad20c34bc6651138d2b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/261534
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>