Check the availability of free command allocator
When testing WebGPU support on Qualcomm, there is a little chance we couldn't get a free allocator. However, it's too hard to reproduce the issue locally. So we use DAWN_CHECK instead of DAWN_ASSERT to see how often we may hit this in the wild.

Bug: 442950558
Change-Id: Ib9636de25d7876d08edfff4b58d4f6270f4c1536
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262254
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Yang Gu <ygu@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>