Rename cts.js to cts.cjs
This is so that it is always interpreted as an CommonJS file.
This issue comes up because there is no package.json in
the build folder so, if there happens to be one above it
it may get interpreted at the wrong type. Changing the
extension fixes that issue.

Note: I tried updating it to use .mjs and modernizing
the code but other code and tools were trying to `require` it.
`require` only loads js/cjs files.

Bug: 446157730
Fixes: 446157730
Change-Id: I573a635f4abaee82eb13efd6d4e09ed02d9a711d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262536
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: Gregg Tavares <gman@chromium.org>