Fix propagation issue in spirv-reader
When getting a Value in the loop header, if the SPIR-V value exists, we
can attempt to get it immediately to force it to propagate if needed.
This fixes an issue where a `break-if` needs to propagate a value up to
the merge block.

Fixed: 444481342
Change-Id: Iac1cbc9a240f67595eefceb9c0bbed390b6e2a78
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/263214
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>