Add gn_v2 Mac/x64/Clang/SwS builders
Adds dawn-mac-x64-builder-rel and dawn-mac-x64-sws-rel. Not tests are
currently enabled, but the same tests as the current mac-rel bot will be
added in a follow-up CL.

Similar to Windows/x64 testers, the machines used for tester builds are
not the standard 2-core Linux GCEs since there are non-Swarmed tests
that need to run. This is expected to be resolved long-term once all
tests are runnable on Swarming.

Bug: 385317083
Change-Id: I623c59f4b1dc4d07a1665abf088c80baf9aba2a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/263037
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@google.com>