Avoid triggering the device lost event during init
If there is a device loss during DeviceBase::Initialize, the device lost
event is triggered and then triggered again by the adapter because the
device failed to initialize. Avoid this case by storing the device lost
event in a "pending" state and moving it to mLostEvent when it's
acceptable to trigger it.

Bug: 406522796
Change-Id: I9880adc46ba1452d94d8f7069f9529a8dedc6802
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/263194
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>