Temporarily target x64 machines for Mac builds
Temporarily updates the gn_v2 Mac builders to only target x64 machines.
This is necessary because the copy of Node that is retrieved depends on
the architecture of the host used for builds. This causes problems if an
ARM64 version is retrieved, as we then try to run it on an x64 machine.

This can be removed once the Node dependency has been updated to
retrieve both x64 and ARM64 versions and users are updated to select the
correct one at runtime.

Bug: 441328362
Change-Id: I0000da2ff3b6af055d45133493c4e58028791be3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/263754
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>