[wgsl] Fix address-of-vector-component check for pointers
WGSL does not support taking the address of a vector component. The
check for this was written prior to the `pointer_composite_access`
language feature, and so was only checking for references. We need to
unwrap pointers too.

Bug: 445793511
Change-Id: Ie9c19cd9c03d7feb59ed01afbe572032aa4b8439
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264134
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: James Price <jrprice@google.com>